+++ src/Portofino.ts
@@ -1,23 +1,33 @@
-import {Action, RootAction} from "./actions/internal";
-import {AxiosInstance} from "axios";
+import { Action, RootAction, LoginAction } from "./actions/internal";
+import { AxiosInstance } from "axios";
 import NooNoo from "./NooNoo";
-import LoginActionManager from "./LoginActionManager";
+
+export * from "./Actions/internal";
 
 export interface PortofinoConfig {
   url?: string;
   axiosInstance?: AxiosInstance;
+
+  enableAuth?: boolean,
   authAction?: string;
+
   crudActionClasses?: string[];
 }
 
 export default class Portofino {
   private static rootAction: RootAction;
-  public static auth: LoginActionManager;
+  public static auth: LoginAction;
 
-  private constructor() {}
+  private constructor() { }
 
-  static connect(config: PortofinoConfig) {
-    const baseURL = config.axiosInstance ? '' : (config.url || '/api');
+  static connect({
+    url = '/api',
+    axiosInstance,
+    enableAuth = true,
+    authAction = 'login',
+    crudActionClasses
+  }: PortofinoConfig) {
+    const baseURL = axiosInstance ? '' : url;
 
     if (baseURL)
       console.debug('[Portofino] Connecting to ', baseURL);
@@ -25,11 +35,13 @@ export default class Portofino {
       console.debug('[Portofino] Connecting via axios instance');
 
     if (Portofino.rootAction)
-      throw new Error("PortofinoJS is already connected!")
+      throw new Error("PortofinoJS is already connected!");
 
-    const noo = NooNoo.create(baseURL, config.axiosInstance);
-    this.rootAction = new RootAction(noo, config.crudActionClasses);
-    this.auth = new LoginActionManager(noo)
+    const noo = NooNoo.create(baseURL, axiosInstance);
+    this.rootAction = new RootAction(noo, crudActionClasses);
+
+    if (enableAuth)
+      this.auth = new LoginAction(noo, authAction, crudActionClasses);
   }
 
   static getAction(name: string): Promise<Action> {